PEAK-27 - Code Cleanup 1.0 / 1.1 - BIG - Core Logic Foundation REFACTORE + FIX + MIGRATION

+++ Application/Source/PeakPerformance.Application/BusinessLogic/Users/Commands/SigninCommand.cs
@@ -22,7 +22,7 @@ public class SigninCommand(SigninDto user) : BaseCommand<AuthorizationDto>
             UserActivityLog userActivityLog = new()
             {
                 UserId = existingUser.Id,
-                ActionTypeId = (int)eActionType.Signin
+                ActionTypeId = eActionType.Signin
             };
 
             await _unitOfWork.UserActivityLogRepository.AddAsync(userActivityLog, cancellationToken);
